Fix MultiGPU double Init with PyN2D2
## Context

The Python API initializes several times deepNets with the Keras Interop. However, it is an impossible behaviour with the MultiGPU.
The objective of this MR is to fix this issue.

## Issue

Some logs about the issue:
```
CUBLAS initialized on device #5
Cuda failure: peer access is already enabled (704) in n2d2/src/DeepNet.cpp:1027 on device #7

### Training ###

# Train Epoch: 0 #
Traceback (most recent call last):
  File "../quantizer_ONNX.py", line 153, in <module>
    x = provider.read_random_batch()
  File "n2d2/python/n2d2/provider.py", line 186, in read_random_batch
    self._deepnet.N2D2().initialize()
RuntimeError: Cuda failure: peer access is already enabled (704) in n2d2/src/DeepNet.cpp:1027 on device #7
```

To speed up the communications among devices, there is a CUDA option to enable peer accesses among devices when it is possible.
The activation of this option happens during the initialization of the deepNet but should only happen once. With the Python API, the initialization is executed more than one time and creates this issue.

## Solution

To fix the problem, the "enable function" has been moved to CudaUtils and should be executed just at the beginning of the program.
Moreover, if the function is called more than one time, a warning will be raised.