Fix BatchNorm fuse with tf2onnx
## Context

Currently, in order to run a training of a nn with the *Keras Interop*, it is necessary to use `tf2onnx` to export the model from TF to N2D2. However, this tool automatically fuse the Batchnorm layers with the Conv ones.

A solution proposed was to use:
```python
if "remove_back_to_back" in tf2onnx.optimizer._get_optimizers():
        tf2onnx.optimizer._get_optimizers().pop("remove_back_to_back")
```

Unfornatunately, these lines introduced a bug with a Reshape issue.

## Issue

The problem happens if a model has two consecutives Reshape layers. These layers are normally removed by `tf2onnx` via *remove_back_to_back* module to produce only one Reshape layer (see https://github.com/onnx/tensorflow-onnx/blob/f64772ce166ea2a0402524d9741b2fb71e5663df/tf2onnx/optimizer/back_to_back_optimizer.py#L257)
```python
    @staticmethod
    @_register_func('Reshape')
    def _optimize_reshape_reshape(g, node, consumer_nodes):
        """remove sequential reshape nodes"""
        if node.type != 'Reshape' or len(consumer_nodes) != 1:
            return []

        node2 = consumer_nodes[0]
        if node2.type != 'Reshape':
            return []

        g.replace_inputs(node2, [node.input[0], node2.input[1]])
        g.remove_node(node.name)
        return []
```

Without *remove_backto_back*, the Reshape optimization doesn't happen.

## Solution

The solution proposed is to only remove the Batchnorm fuse in the optimizer instead of deleting the whole BackToBackOptimizer module. So the following line will replace the former one
```python
if "remove_back_to_back" in tf2onnx.optimizer._get_optimizers():
        tf2onnx.optimizer.back_to_back_optimizer._func_map.pop(('Conv', 'BatchNormalization'))
```

Then these lines are added to a context so the Batchnorm fuse is re-included in the optimizer at the exit of the context.

## Unplanned fix

It is now possible to export your dataset with the Python API on embedded platforms. This PR also resolves a problem where it was not possible to export the dataset because the stimuliProvider of the target in the deepNet was not initialized.