Fix CMake and compilation warnings
## Context

The objective of this PR is to resolve compilation warnings as well as a `CMake` warning.

Compilation warnings are minor (see the commits to understand).

The `CMake` warning happens when N2D2 is adding the Binding functions to the compilation. The warning doesn't occur when N2D2 is used in an extra module but happens when N2D2 is not used in an extra module.

## Solution

The solution was to include a variable in the extra module `CMakeLists.txt` of the extra module to import the N2D2 Python binding and performs a test to know if the variable is defined in the `CMakeLists.txt` of N2D2.

So in the development of the upcoming extra modules using N2D2, don't forget to include
```cmake
set(N2D2_EXTRAMODULE 1)
```
at the beginning of the `CMakeLists.txt`.