Add Learned Stepsize Quantization (LSQ) method
## Context

The objective of this PR is to propose the [LSQ method](https://arxiv.org/pdf/1902.08153.pdf) in N2D2 in order to quantize neural networks with a precision lower than 8-bit.

## Current program behaviour

No Quantization Aware-Training (QAT) is implemented.

## Behaviour after modifications

The user will be able to use LSQ to quantize his/her model with the INI or Python interface. <br>
Several examples are proposed to use LSQ in `python/examples`.

## Detailed major modifications

This PR introduces a new directory `Quantizer/QAT`. The aim of this directory is to contain all QAT methods used in N2D2. <br>
This directory is separated into 3 folders:
* *Activation*: methods to quantize activations
* *Cell*: methods to quantize the cells (essentially the parameters)
* *Kernel*: kernels to quantize on CPU and on CUDA