Hotfix cpp export
## Context

The current `CPP export` presents multiple working issues, due to several implementation problems of the quantization tools into the CPP export. The goal of this PR is to propose the former version of the CPP export, tested and stable and a dedicated space for quantified export to improve it without disrupting the rest of the exports.

## Issue

It was impossible to compile an CPP export after generating it because some conflicts occured.

```
./include/Network.hpp:2514:45: error: no matching function for call to ‘N2D2::Network::macsOnRange<64>(signed char*, const data<8>* __restrict__, int&) const’
                     macsOnRange<NB_CHANNELS>(
                     ~~~~~~~~~~~~~~~~~~~~~~~~^
                                 (Input_T*)((uint8_t*)inputs + iOffsetInRange),
                                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                                 weights + wOffset + ix * W_BYTES,
                                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                                 weightedSum);
```

There was also a conflict with the *Memory Manager* if one wanted to generate an export in *Float32*. In that case, it allocated 4 times too much memory than necessary.

Another problem has been detected when one tries to remove the last cell of a network. The associated target doesn't point to the new last cell of the network if this case happens.

## Solution

This PR proposes multiple fixes and updates:
* Added the former CPP export and the current one is moved to `CPP_Quantization`
* The Memory Manager has been fixed for the *Float32* export
* DeepNet::removeCell now updates the cell pointed by the Target if the removed cell is the last one
* Multiple files in the CPP export have been improved like the `Makefile` or `cpp_utils.cpp`.

## Additional information

The path to the binaries that can be generated after compiling the CPP export has been changed:
* Before the PR, one has to run `./run_export`
* After the PR, he/she has to run `./bin/run_export`