Add SAT QAT module
## Context

With the current version of N2D2, it is only possible to use the Quantization Aware-Training module with the *LSQ* module. <br>
The aim of this PR is to add another QAT module, known as [SAT](https://arxiv.org/pdf/1912.10207.pdf) in N2D2 in order to quantize neural networks with a precision lower than 8-bit.

## Detailed major modifications

The user will be able to use SAT to quantize his/her model with the INI or Python interface. <br>
Several examples are proposed to use SAT in `python/examples`.

Moreover, the PR introduces a new QAT block called "Optimization/DeepNetQAT". <br>
This block is mainly used to optimize the graph of a quantized model before exporting it.

## Unplanned changes

Several issues have been fixed in this branch:
- a fix for a segmentation-fault event when the user used the *-learn-epoch* option on small datasets
- a fix in some exports where the namespace wasn't changed from `N2D2` to `N2D2_Export`