Pruning mode adding
## Context

Among existing compression techniques, the pruning method has proved that it is a viable method to compress the memory footprint of a model. The aim of this PR is to add a pruning mode in N2D2 to increase the number of compression tools present in the framework.

## Behaviour after modifications

It is possible to prune a model with the `PruneCell`.
See more details in the Pruning documentation section to use it.

## Unplanned changes

Fixed in this branch an issue with the Github CI. Indeed since November 2022, the Ubuntu version in the Github servers has changed from 20.04 to 22.04. Now the Github CI of N2D2 will use `ubuntu-20.04` again.