Export STM32
## Context

The aim of this PR is to add the N2D2 export for STM32 in open-source. <br>
You can export now your model in STM32 with the same commands than the CPP export.

Please go to the `export/CPP_STM32` folder to see how to use your export on board !

## Unplanned changes

This PR also fixed some issues for 32-bit processors.